+++ src/main/java/com/sparta/newsfeed_project/domain/common/config/JwtConfig.java
@@ -2,7 +2,7 @@ package com.sparta.newsfeed_project.domain.common.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import com.sparta.newsfeed_project.domain.common.jwt.JwtUtil;
+import com.sparta.newsfeed_project.domain.common.util.JwtUtil;
 
 /**
  * JWT 토큰 유틸 설정을 담당하는 클래스입니다.
