+++ src/main/java/com/sparta/newsfeed_project/domain/common/util/PasswordEncoder.java
@@ -1,4 +1,4 @@
-package com.sparta.newsfeed_project.domain.common.jwt;
+package com.sparta.newsfeed_project.domain.common.util;
 
 import at.favre.lib.crypto.bcrypt.BCrypt;
